[impl] restructure the modules for the #[bitfield] macro
